Renames Controller to Controls
Adds Controls to app.js
